Get closer to supporting Confluence 4.0, partially fixes JENKINS-11276
The Confluence 4.0 update includes a new SOAP endpoint (v2), and the
v1 endpoint's getPage() method now results in a remote error --
suggesting instead to use getPageSummary().  The file attachment
functionality did not rely on the wiki content being returned in
getPage(), so we now use getPageSummary() for doing file attachments.

Confluence 4.0 also introduces a new page content storage format,
which means wiki markup is no longer supported.  That means the page
editing feature of the plugin will need to be updated to work with the
v2 API's XHTML storage format.  For now, we're just fixing the file
attachment bug, and gracefully degrading on page content editing.