JENKINS-11276: A hack to get Confluence v4.0 markup edits working.
In Confluence 4.0, you have to use the v2 SOAP endpoint in order
to fetch the page content, and then the content can only be
edited in an XML format.

The plugin is not very intelligent at the moment, as far as
understanding the new <ac:macro/>-style XML tags, and being able
to simplify the token filtering process. Instead, you'll need
to enter the tokens in their X(HT)ML format, which can be found
by going to Tools > View Storage Format from a Confluence page.
Also note that these tokens can simply be visible tokens within
the markup.